PrototypesCriticisms

Documentation for PrototypesCriticisms.

PrototypesCriticisms.importances

PrototypesCriticisms.importances

PrototypesCriticisms.importances

PrototypesCriticisms.prototypes_criticisms

PrototypesCriticisms.prototypes_criticisms

PrototypesCriticisms.importances — Method

importances(observed_ranking, rankings)

Compute the importances for every data instance.

Arguments

observed_ranking::Vector{Int}: the originally observed ranking.

rankings::Matrix{Int}: the rankings per data feature.

PrototypesCriticisms.importances — Method

importances(observed_ranking, ranking)

Like importances, but provide a single ranking for one data feature only.

PrototypesCriticisms.importances — Method

importances(observed_ranking, rankings)

Like importances, but provide a vector of rankings instead of a matrix.

PrototypesCriticisms.prototypes_criticisms — Method

prototypes_criticisms(observed_ranking, rankings, p, c)

Find prototypes and criticisms.

Arguments

observed_ranking::Vector{Int}: the observed ranking.

rankings::Matrix{Int}: the rankings per data feature.

p::Int: the number of protoypes to find.

c::Int: the number of criticisms to find.

Notes

Each ranking contains the data instance identifiers, i.e., every ranking is a mapping of indices to data instances, e.g., observed_ranking[3] = "data instance at rank 3"

PrototypesCriticisms.prototypes_criticisms — Method

prototypes_criticisms(observed_ranking, rankings, p, c)

Like prototypes_criticisms, but provide a vector of rankings instead of a matrix.